requirements.txt
# Core dependencies
dnspython>=2.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
rich>=13.7.0
urllib3>=2.0.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0 
